carregar_energia_api/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.57.0
django-cors-headers==4.2.0
django-filter==1.1.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.0
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
mysqlclient==2.2.0
unstructured-client==0.23.7
firebase-admin == 6.6.0